pyproject.toml
[tool.poetry]
name = "users-service"
version = "0.1.0"
description = ""
authors = ["Nickolay <rainbowdash593@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.1"
pydantic = "^1.10.1"
fastapi = "^0.81.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
SQLAlchemy = "^1.4.40"
python-dotenv = "^0.21.0"
psycopg2 = "^2.9.3"
alembic = "^1.8.1"
databases = {extras = ["postgresql"], version = "^0.6.1"}
greenlet = "^1.1.3"
passlib = "^1.7.4"
python-jose = "^3.3.0"
bcrypt = "^4.0.0"
python-multipart = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
